simulated_sasca/requirements.txt
attrs==23.2.0
cattrs==23.2.3
contourpy==1.2.1
cycler==0.12.1
docstring-to-markdown==0.15
exceptiongroup==1.2.1
fonttools==4.53.0
jedi==0.19.1
kiwisolver==1.4.5
loguru==0.7.2
lsprotocol==2023.0.1
matplotlib==3.9.0
matplotlib-backend-wezterm==2.1.2
numpy==1.26.4
packaging==24.1
parso==0.8.4
pillow==10.4.0
pluggy==1.5.0
py-cpuinfo==9.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-lsp-jsonrpc==1.1.2
python-lsp-ruff==2.2.1
python-lsp-server==1.11.0
ruff==0.4.10
scalib==0.5.8
scipy==1.14.0
six==1.16.0
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
ujson==5.10.0
